sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-sentient"
version = "0.3.0"
description = "Claude Sentient SDK - Autonomous Development Orchestration Layer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Claude Sentient Contributors" }
]
keywords = [
    "claude",
    "ai",
    "autonomous",
    "development",
    "orchestration",
    "sdk",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
sdk = [
    "claude-agent-sdk>=0.1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "pyright>=1.1",
]

[project.scripts]
claude-sentient = "claude_sentient.orchestrator:main"
cs = "claude_sentient.orchestrator:main"

[project.urls]
Homepage = "https://github.com/anthropics/claude-sentient"
Documentation = "https://github.com/anthropics/claude-sentient#readme"
Repository = "https://github.com/anthropics/claude-sentient"

[tool.hatch.build.targets.wheel]
packages = ["claude_sentient"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
